+++ internal/test/e2e/framework/node.go
