+++ internal/test/e2e/framework/stats.go
