+++ vkubelet/env_internal_test.go
@@ -11,7 +11,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	testutil "github.com/virtual-kubelet/virtual-kubelet/test/util"
+	testutil "github.com/virtual-kubelet/virtual-kubelet/internal/test/util"
 )
 
 const (
