+++ vkubelet/pod_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/virtual-kubelet/virtual-kubelet/errdefs"
-	testutil "github.com/virtual-kubelet/virtual-kubelet/test/util"
+	testutil "github.com/virtual-kubelet/virtual-kubelet/internal/test/util"
 	"gotest.tools/assert"
 	is "gotest.tools/assert/cmp"
 	corev1 "k8s.io/api/core/v1"
